Plane Fight added (Not finished)
Alright so the first part of the plane fight has been added.

Right now the whole intro part is all scripted and so are 2 of the
plane's attacks.
Currently the only way to do damage to the plane is by shooting it, this
will be changed to acrobatic attacks on the cockpit eventually.
If you somehow manage to beat the plane (Good luck) nothing happens yet

Some other bugs and changes have been made im sure too

+++ src/Entities/Enemy.java
@@ -48,7 +48,7 @@ public abstract class Enemy extends MapObject
 	//slowdown (when player picks up powerup)
 	protected float slowDown;
 	
-	public Enemy(int x, int y, TileMap tm, Player player)
+	public Enemy(double x, double y, TileMap tm, Player player)
 	{
 		super(tm);
 		this.x = x;
